requirements.txt
streamlit>=1.39.0
streamlit-mic-recorder>=0.0.3
groq>=0.4.2
openai-whisper>=20231117
python-dotenv>=1.0.1
pydantic>=2.4.0
matplotlib>=3.7.0
numpy>=1.24.0
